Add --destination-file command line option
This option can be specified multiple lines and expects to receive a
line delimited file containing a list of destination tags for the image
push.

If both --destination and --destination-file are specified then the
values are combined and deduplicated.

This feature allows for easier use in CI environments where a list of
tags to push may be calculated elsewhere.